tweaks

+++ 12.3_BP_calc_tx.sh
@@ -15,18 +15,17 @@ if [ ! -f "$ADA_USB_MNT/hot-env/deleg.cert" ]; then
 	exit 1
 fi
 
+if [ ! -f "$ADA_USB_MNT/hot-env/pool.cert" ]; then
+	echo "no pool.cert!"
+	exit 1
+fi
+
 if [ ! -f "$ADA_USB_MNT/hot-env/payment.addr" ]; then
 	echo "no payment.addr!"
 	exit 1
 fi
 
 cp $ADA_USB_MNT/hot-env/payment.addr hot-env/
-mkdir -p pool-keys
-
-cardano-cli query protocol-parameters \
-    --mainnet \
-    --allegra-era \
-    --out-file pool-keys/params.json
 
 currentSlot=$(cardano-cli query tip --mainnet | jq -r '.slotNo')
 echo Current Slot: $currentSlot
@@ -34,10 +33,10 @@ echo Current Slot: $currentSlot
 cardano-cli query utxo \
     --address $(cat hot-env/payment.addr) \
     --allegra-era \
-    --mainnet > pool-keys/fullUtxo.out
+    --mainnet > fullUtxo.out
 
-tail -n +3 pool-keys/fullUtxo.out | sort -k3 -nr > pool-keys/balance.out
-cat pool-keys/balance.out
+tail -n +3 fullUtxo.out | sort -k3 -nr > balance.out
+cat balance.out
 
 tx_in=""
 total_balance=0
@@ -49,14 +48,21 @@ while read -r utxo; do
     echo TxHash: ${in_addr}#${idx}
     echo ADA: ${utxo_balance}
     tx_in="${tx_in} --tx-in ${in_addr}#${idx}"
-done < pool-keys/balance.out
+done < balance.out
 
-txcnt=$(cat pool-keys/balance.out | wc -l)
-rm pool-keys/balance.out
+txcnt=$(cat balance.out | wc -l)
+rm balance.out
 echo Total ADA balance: ${total_balance}
 echo Number of UTXOs: ${txcnt}
 
-poolDeposit=$(cat pool-keys/params.json | jq -r '.poolDeposit')
+# for testing
+if [ -z "$tx_in" ]; then
+	echo "NOTE: 0 balance - using test account"
+	tx_in="--tx-in aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa#0"
+	total_balance="5000000000"
+fi
+
+poolDeposit=$(cat pool-files/params.json | jq -r '.poolDeposit')
 echo poolDeposit: $poolDeposit
 
 cardano-cli transaction build-raw \
@@ -64,25 +70,21 @@ cardano-cli transaction build-raw \
     --tx-out $(cat hot-env/payment.addr)+0 \
     --invalid-hereafter $(( ${currentSlot} + 10000)) \
     --fee 0 \
-    --out-file pool-keys/tx.tmp \
+    --out-file tx.tmp \
     --allegra-era \
     --certificate $ADA_USB_MNT/hot-env/stake.cert
 
 fee=$(cardano-cli transaction calculate-min-fee \
-    --tx-body-file pool-keys/tx.tmp \
+    --tx-body-file tx.tmp \
     --tx-in-count ${txcnt} \
     --tx-out-count 1 \
     --mainnet \
     --witness-count 3 \
     --byron-witness-count 0 \
-    --protocol-params-file pool-keys/params.json | awk '{ print $1 }')
+    --protocol-params-file pool-files/params.json | awk '{ print $1 }')
 echo fee: $fee
 
-# for changing after initial submission
-txOut=$((${total_balance}-${fee}))
-
-# initial transaction
-#txOut=$((${total_balance}-${poolDeposit}-${fee}))
+txOut=$((${total_balance}-${poolDeposit}-${fee}))
 echo txOut: ${txOut}
 
 cardano-cli transaction build-raw \
@@ -95,13 +97,8 @@ cardano-cli transaction build-raw \
     --allegra-era \
     --out-file $ADA_USB_MNT/offline-files/tx.raw
 
-echo "tx.raw"
-echo "==================================="
+echo "--- tx.raw"
 cat $ADA_USB_MNT/offline-files/tx.raw
 
-rm pool-keys/fullUtxo.out
-rm pool-keys/tx.tmp
-rm $ADA_USB_MNT/hot-env/pool.cert
-rm $ADA_USB_MNT/hot-env/stake.cert
-rm $ADA_USB_MNT/hot-env/deleg.cert
-
+rm fullUtxo.out
+rm tx.tmp
